backend/requirements.txt
Flask
Flask_SQLAlchemy
Authlib
PyJWT
flasgger
python-dotenv
requests